requirements.txt
# 核心依赖
flask==2.3.2
neo4j==5.11.0
numpy==1.24.3
torch==2.0.1
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2

# 数据处理
tqdm==4.65.0
python-dotenv==1.0.0

# 开发工具
pytest==7.4.0
black==23.7.0
flake8>=7.0.1